Added connect left end; 3699 full-length for SRR097897

+++ src/tpl.hpp
@@ -144,7 +144,7 @@ extern "C" {
 	void mv_paired_reads_to_tried(hash_table *ht, tpl *t);
 	void mv_paired_reads_back(tpl *t);
 	void mv_reads_to_main_tpl(tpl *t, tpl *b, int from);
-	int pairs_connect_left_locus(bwa_seq_t *seqs, tpl *t, tpl *b, int locus, int from, float *n_unpaired);
+	float pairs_connect_left_locus(bwa_seq_t *seqs, tpl *t, tpl *b, int locus, int from, float *n_unpaired);
 
 #ifdef __cplusplus
 }
